// client/src/core/color/types.ts
export type TColorString = 'red' | 'blue' | 'green' | 'yellow' | 'purple';

export type TColorHex =
  | '#e32990'
  | '#fff110'
  | '#4cba35'
  | '#28a3dc'
  | '#9719ff';

export type TColorRbg =
  | 'rgba(227,41,144,0.2)'
  | 'rgba(255,241,16,0.2)'
  | 'rgba(76,186,53,0.2)'
  | 'rgba(40,163,220,0.2)'
  | 'rgba(151,25,255,0.2)';